weather_project/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.2.1
cycler==0.12.1
DateTime==5.5
Django==5.1.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
et-xmlfile==1.1.0
fonttools==4.53.1
idna==3.10
kiwisolver==1.4.5
matplotlib==3.9.2
numpy==2.1.0
openpyxl==3.1.5
packaging==24.1
panda==0.3.1
pandas==2.2.2
patsy==0.5.6
pillow==10.4.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
scipy==1.14.1
seaborn==0.13.2
setuptools==75.6.0
six==1.16.0
sqlparse==0.5.3
statsmodels==0.14.4
tzdata==2024.1
urllib3==2.2.3
zope.interface==7.1.1
